refactor(web): enforce architecture rules on shared and repoint deep imports
