+++ apps/web/src/main.tsx
@@ -3,7 +3,7 @@ import { createRoot } from 'react-dom/client';
 import { QueryClientProvider } from '@tanstack/react-query';
 import { queryClient } from '@/lib/queryClient';
 import App from './App';
-import ErrorBoundary from '@/components/shared/ErrorBoundary';
+import ErrorBoundary from '@/components/shared/ErrorBoundary/ErrorBoundary';
 import { Toaster } from '@/components/ui/sonner';
 import { TooltipProvider } from '@/components/ui/tooltip';
 import { initSentryRenderer, captureException } from '@/lib/sentry';
